немного рефактора
