+++ run_tests.py
@@ -13,7 +13,7 @@ if __name__ == '__main__':
     suite = unittest.TestSuite((
         unittest.makeSuite(PositiveLoginTest),
         unittest.makeSuite(NegativeLoginTest),
-        unittest.makeSuite(MainPageTest)
+        unittest.makeSuite(MainPageTest),
     ))
     result = unittest.TextTestRunner().run(suite)
     sys.exit(not result.wasSuccessful())
