+++ tests/__init__.py
@@ -1,0 +1,3 @@
+from tests.base import Page, Component
+
+__all__ = (Page, Component)
